LPS-76524 Move project template customizers to internal package
