+++ modules/sdk/project-templates/project-templates-fragment/src/main/resources/META-INF/services/com.liferay.project.templates.ProjectTemplateCustomizer
@@ -1,1 +1,1 @@
-com.liferay.project.templates.fragment.FragmentProjectTemplateCustomizer
+com.liferay.project.templates.fragment.internal.FragmentProjectTemplateCustomizer
