+++ modules/sdk/project-templates/project-templates-npm-angular-portlet/src/main/resources/META-INF/services/com.liferay.project.templates.ProjectTemplateCustomizer
@@ -1,1 +1,1 @@
-com.liferay.project.templates.npm.angular.portlet.NpmAngularPortletProjectTemplateCustomizer
+com.liferay.project.templates.npm.angular.portlet.internal.NpmAngularPortletProjectTemplateCustomizer
