+++ modules/sdk/project-templates/project-templates-npm-billboardjs-portlet/src/main/resources/META-INF/services/com.liferay.project.templates.ProjectTemplateCustomizer
@@ -1,1 +1,1 @@
-com.liferay.project.templates.npm.billboardjs.portlet.NpmBillboardjsPortletProjectTemplateCustomizer
+com.liferay.project.templates.npm.billboardjs.portlet.internal.NpmBillboardjsPortletProjectTemplateCustomizer
