+++ modules/sdk/project-templates/project-templates-npm-isomorphic-portlet/src/main/resources/META-INF/services/com.liferay.project.templates.ProjectTemplateCustomizer
@@ -1,1 +1,1 @@
-com.liferay.project.templates.npm.isomorphic.portlet.NpmIsomorphicPortletProjectTemplateCustomizer
+com.liferay.project.templates.npm.isomorphic.portlet.internal.NpmIsomorphicPortletProjectTemplateCustomizer
