+++ modules/sdk/project-templates/project-templates-npm-jquery-portlet/src/main/resources/META-INF/services/com.liferay.project.templates.ProjectTemplateCustomizer
@@ -1,1 +1,1 @@
-com.liferay.project.templates.npm.jquery.portlet.NpmJqueryPortletProjectTemplateCustomizer
+com.liferay.project.templates.npm.jquery.portlet.internal.NpmJqueryPortletProjectTemplateCustomizer
