+++ modules/sdk/project-templates/project-templates-npm-metaljs-portlet/src/main/resources/META-INF/services/com.liferay.project.templates.ProjectTemplateCustomizer
@@ -1,1 +1,1 @@
-com.liferay.project.templates.npm.metaljs.portlet.NpmMetaljsPortletProjectTemplateCustomizer
+com.liferay.project.templates.npm.metaljs.portlet.internal.NpmMetaljsPortletProjectTemplateCustomizer
