+++ modules/sdk/project-templates/project-templates-npm-portlet/src/main/resources/META-INF/services/com.liferay.project.templates.ProjectTemplateCustomizer
@@ -1,1 +1,1 @@
-com.liferay.project.templates.npm.portlet.NpmPortletProjectTemplateCustomizer
+com.liferay.project.templates.npm.portlet.internal.NpmPortletProjectTemplateCustomizer
