+++ modules/sdk/project-templates/project-templates-npm-react-portlet/src/main/resources/META-INF/services/com.liferay.project.templates.ProjectTemplateCustomizer
@@ -1,1 +1,1 @@
-com.liferay.project.templates.npm.react.portlet.NpmReactPortletProjectTemplateCustomizer
+com.liferay.project.templates.npm.react.portlet.internal.NpmReactPortletProjectTemplateCustomizer
