+++ modules/sdk/project-templates/project-templates-npm-vuejs-portlet/src/main/resources/META-INF/services/com.liferay.project.templates.ProjectTemplateCustomizer
@@ -1,1 +1,1 @@
-com.liferay.project.templates.npm.vuejs.portlet.NpmVuejsPortletProjectTemplateCustomizer
+com.liferay.project.templates.npm.vuejs.portlet.internal.NpmVuejsPortletProjectTemplateCustomizer
