+++ modules/sdk/project-templates/project-templates-service-builder/src/main/resources/META-INF/services/com.liferay.project.templates.ProjectTemplateCustomizer
@@ -1,1 +1,1 @@
-com.liferay.project.templates.service.builder.ServiceBuilderProjectTemplateCustomizer
+com.liferay.project.templates.service.builder.internal.ServiceBuilderProjectTemplateCustomizer
