+++ modules/sdk/project-templates/project-templates-service-wrapper/src/main/resources/META-INF/services/com.liferay.project.templates.ProjectTemplateCustomizer
@@ -1,1 +1,1 @@
-com.liferay.project.templates.service.wrapper.ServiceWrapperProjectTemplateCustomizer
+com.liferay.project.templates.service.wrapper.internal.ServiceWrapperProjectTemplateCustomizer
