+++ modules/sdk/project-templates/project-templates-service/src/main/resources/META-INF/services/com.liferay.project.templates.ProjectTemplateCustomizer
@@ -1,1 +1,1 @@
-com.liferay.project.templates.service.ServiceProjectTemplateCustomizer
+com.liferay.project.templates.service.internal.ServiceProjectTemplateCustomizer
