+++ modules/sdk/project-templates/project-templates-soy-portlet/src/main/resources/META-INF/services/com.liferay.project.templates.ProjectTemplateCustomizer
@@ -1,1 +1,1 @@
-com.liferay.project.templates.soy.portlet.SoyPortletProjectTemplateCustomizer
+com.liferay.project.templates.soy.portlet.internal.SoyPortletProjectTemplateCustomizer
