+++ modules/sdk/project-templates/project-templates-theme-contributor/src/main/resources/META-INF/services/com.liferay.project.templates.ProjectTemplateCustomizer
@@ -1,1 +1,1 @@
-com.liferay.project.templates.theme.contributor.ThemeContributorProjectTemplateCustomizer
+com.liferay.project.templates.theme.contributor.internal.ThemeContributorProjectTemplateCustomizer
